+++ src/Domain/Module/Nginx/VirtualHost.hs
@@ -6,12 +6,14 @@ module Domain.Module.Nginx.VirtualHost (
 
 import Data.Aeson
 import Data.Text (Text)
+import Policy (Policy)
 
 data NginxVirtualHost = NginxVirtualHost
     { vhostDomain :: Text
     , vhostHttpPort :: Maybe Int
     , vhostHttpsPort :: Maybe Int
     , vhostProxyPass :: Maybe Text
+    , vhostPolicies :: [Policy]
     }
     deriving (Eq, Show)
 
@@ -22,6 +24,7 @@ instance ToJSON NginxVirtualHost where
                 <> maybe [] (\p -> ["http_port" .= p]) (vhostHttpPort vh)
                 <> maybe [] (\p -> ["https_port" .= p]) (vhostHttpsPort vh)
                 <> maybe [] (\u -> ["proxy_pass" .= u]) (vhostProxyPass vh)
+                <> (if null (vhostPolicies vh) then [] else ["policies" .= vhostPolicies vh])
 
 instance FromJSON NginxVirtualHost where
     parseJSON = withObject "NginxVirtualHost" $ \o ->
@@ -30,3 +33,4 @@ instance FromJSON NginxVirtualHost where
             <*> o .:? "http_port"
             <*> o .:? "https_port"
             <*> o .:? "proxy_pass"
+            <*> o .:? "policies" .!= []